requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyJWT==2.9.0
requests==2.32.3
python-dotenv==1.0.1
langchain-ollama==0.2.0
langchain-core==0.3.15
langchain==0.3.7
